+++ src/components/action-popover/action-popover.d.ts
@@ -1,6 +1,7 @@
 import * as React from 'react';
+import { MarginSpacingProps } from "../../utils/helpers/options-helper";
 
-export interface ActionPopoverProps {
+export interface ActionPopoverProps extends MarginSpacingProps {
   id?: string;
   onOpen?: () => void;
   onClose?: () => void;